requirements.txt
flask==3.1.2
requests==2.32.3
python-dotenv==1.0.1
wechatpy==1.8.18
